requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pygments==2.17.2
rich==13.7.0
Werkzeug==3.0.1
virustotal-api==1.1.11 
yara-python==4.3.1 
requests==2.31.0